(PC-25475) refacto(OfferBookingButton): move to offerv2
